move WorkQueue out of p4p.rpc
