+++ src/p4p/rpc.py
@@ -3,12 +3,6 @@ import logging, inspect
 from functools import wraps, partial
 _log = logging.getLogger(__name__)
 
-try:
-    from Queue import Queue, Full, Empty
-except ImportError:
-    from queue import Queue, Full, Empty
-from threading import Thread
-
 __all__ = [
     'rpc',
     'rpccall',
@@ -21,6 +15,7 @@ __all__ = [
 from .wrapper import Value, Type
 from .nt import NTURI
 from .server import SharedPV, DynamicProvider
+from .util import WorkQueue
 
 def rpc(rtype=None):
     """Decorator marks a proxy method for export.
@@ -70,41 +65,6 @@ def rpccall(pvname, rtype=None, request=None):
 
 class RemoteError(RuntimeError):
     "Throw with an error message which will be passed back to the caller"
-
-class WorkQueue(object):
-    _stopit = object()
-    def __init__(self, maxsize=5):
-        self._Q = Queue(maxsize=maxsize)
-    def push(self, callable):
-        self._Q.put_nowait(callable) # throws Queue.Full
-    def push_wait(self, callable):
-        self._Q.put(callable)
-    def interrupt(self):
-        """Break one call to handle()
-
-        eg. Call N times to break N threads.
-
-        This call blocks if the queue is full.
-        """
-        self._Q.put(self._stopit)
-    def handle(self):
-        """Process queued work until interrupt() is called
-        """
-        while True:
-            # TODO: Queue.get() (and anything using thread.allocate_lock
-            #       ignores signals :(  so timeout periodically to allow delivery
-            try:
-                callable = self._Q.get(True, 1.0)
-            except Empty:
-                continue # retry on timeout
-            try:
-                if callable is self._stopit:
-                    break
-                callable()
-            except:
-                _log.exception("Error from WorkQueue")
-            finally:
-                self._Q.task_done()
 
 class RPCDispatcherBase(DynamicProvider):
     # wrapper to use for request Structures
